renderöidään adminpage vain jos kirjautunut user === admin

+++ frontend/aikatauluttaja/src/mui_components/AdminPage.js
@@ -40,15 +40,23 @@ class AdminPage extends Component {
 
     
   render() {
+
+    if(this.props.user !== null && this.props.user.username === "admin"){
         return (
-        <div>
-            <Button onClick={() => this.deleteAllGoals()} style={{marginLeft: 20, marginBottom: 15, marginTop: 10}} size="small" variant="contained" color="secondary" type="submit">Delete all Goals</Button>
-            <Button onClick={() => this.deleteAllUsers()} style={{marginLeft: 20, marginBottom: 15, marginTop: 10}} size="small" variant="contained" color="secondary" type="submit">Delete all Users</Button>
-            <Button onClick={() => this.deleteAllCourses()} style={{marginLeft: 20, marginBottom: 15, marginTop: 10}} size="small" variant="contained" color="secondary" type="submit">Delete all Courses</Button>
-            <br></br>
-            <Button onClick={() => this.deleteEverything()} style={{marginLeft: 20, marginBottom: 15, marginTop: 10}} size="small" variant="contained" color="secondary" type="submit">DELETE EVERYTHING</Button>
-          </div>
-        )
+            <div>
+                <Button onClick={() => this.deleteAllGoals()} style={{marginLeft: 20, marginBottom: 15, marginTop: 10}} size="small" variant="contained" color="secondary" type="submit">Delete all Goals</Button>
+                <Button onClick={() => this.deleteAllUsers()} style={{marginLeft: 20, marginBottom: 15, marginTop: 10}} size="small" variant="contained" color="secondary" type="submit">Delete all Users</Button>
+                <Button onClick={() => this.deleteAllCourses()} style={{marginLeft: 20, marginBottom: 15, marginTop: 10}} size="small" variant="contained" color="secondary" type="submit">Delete all Courses</Button>
+                <br></br>
+                <Button onClick={() => this.deleteEverything()} style={{marginLeft: 20, marginBottom: 15, marginTop: 10}} size="small" variant="contained" color="secondary" type="submit">DELETE EVERYTHING</Button>
+              </div>
+            )
+    } else {
+        return null
+    }
+
+
+
     }
